requirements.txt
fastapi~=0.110.0
python-dotenv~=1.0.1
torch~=2.2.0
transformers~=4.39.1
pypdf~=4.1.0
uvicorn~=0.29.0